update contacts when personal record exists

+++ address_book_ui/src/app/interfaces/payloads/ContactDetailsCreateBody.ts
@@ -1,5 +1,5 @@
 export interface ContactDetailsCreateBody {
-    recordId: BigInt | null,
+    recordId: BigInteger | null,
     type: string,
     value: string
 }
